teams_agent/requirements.txt
aiohttp>=3.9
python-dotenv>=1.0
msal>=1.28
microsoft-agents-activity>=0.9.0
microsoft-agents-copilotstudio-client>=0.9.0
microsoft-agents-hosting-aiohttp>=0.9.0
microsoft-agents-hosting-teams>=0.9.0
microsoft-agents-authentication-msal>=0.9.0
